requirements.txt
# Food Quality Analyzer - Production Requirements

# Core Framework
streamlit==1.28.1
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL support
sqlite3  # Built-in SQLite support

# AI and ML
langchain-core==0.1.52
langchain-groq==0.1.5
openai==1.3.7
groq==0.4.1

# Image Processing and OCR
pillow==10.0.1
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
numpy==1.24.3

# Data Processing
pandas==2.0.3
pydantic==2.5.0
python-multipart==0.0.6

# Web and API
requests==2.31.0
httpx==0.25.2
fastapi-cors==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and Logging
sentry-sdk==1.38.0
structlog==23.2.0
prometheus-client==0.19.0

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI
factory-boy==3.3.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Production
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
flower==2.0.1

# System monitoring
psutil==5.9.6

# Security
cryptography==41.0.8
bcrypt==4.1.2

# Cloud deployment
boto3==1.34.0  # AWS
google-cloud-storage==2.10.0  # GCP
azure-storage-blob==12.19.0  # Azure

# Docker health checks
docker==6.1.3